requirements.txt
tensorflow_probability==0.21.0
tensorflow_addons==0.21.0
tensorflow_io==0.34.0
click==8.1.7
environs==10.3.0
pandas==2.0.3
geopandas==0.13.2
matplotlib==3.7.2
graphviz==0.20.1
pydot==2.0.0
